GH-3406 Order of optimisers may matter
More tests cases.
Fix copyright year.
